[pack installer] various tweaks from PR review

+++ src/cli_plugin/remove/remove.js
@@ -1,19 +1,23 @@
-import fs from 'fs';
+import { statSync } from 'fs';
 import rimraf from 'rimraf';
 
 export default function remove(settings, logger) {
   try {
+    let stat;
     try {
-      fs.statSync(settings.pluginPath);
+      stat = statSync(settings.pluginPath);
     } catch (e) {
-      logger.log(`Plugin ${settings.plugin} is not installed`);
-      return;
+      throw new Error(`Plugin [${settings.plugin}] is not installed`);
+    }
+
+    if (!stat.isDirectory()) {
+      throw new Error(`[${settings.plugin}] is not a plugin`);
     }
 
     logger.log(`Removing ${settings.plugin}...`);
     rimraf.sync(settings.pluginPath);
   } catch (err) {
-    logger.error(`Unable to remove plugin "${settings.plugin}" because of error: "${err.message}"`);
+    logger.error(`Unable to remove plugin because of error: "${err.message}"`);
     process.exit(74); // eslint-disable-line no-process-exit
   }
 }
